refactor: update tests structure
